[FIX] BestBoard 인기 게시글 컴포넌트에 포함시킨거 제거
